requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
distlib==0.3.3
Django==3.2.9
django-bootstrap4==21.1
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-ftpserver==0.7.0
django-grappelli==2.15.3
django-storages==1.12.3
django-suit==2.0a1
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
filelock==3.3.2
future==0.18.2
html5lib==1.1
humanize==3.13.1
idna==3.3
Markdown==3.3.4
MouseInfo==0.1.3
mysqlclient==2.1.0
pafy==0.5.5
Pillow==8.4.0
platformdirs==2.4.0
PyAutoGUI==0.9.53
pyftpdlib==1.5.6
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.4
PyPDF2==1.26.0
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyRect==0.1.4
PyScreeze==0.1.28
python-bidi==0.4.2
pytweening==1.0.4
pytz==2021.3
reportlab==3.6.2
requests==2.26.0
six==1.16.0
soupsieve==2.3
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
virtualenv==20.13.1
webencodings==0.5.1
xhtml2pdf==0.2.5
youtube-dl==2021.6.6
